+++ lib/repository.c
@@ -592,7 +592,7 @@ got_repo_open(struct got_repository **repop, const char *path,
 	struct got_repository *repo = NULL;
 	const struct got_error *err = NULL;
 	char *abspath, *repo_path = NULL;
-	int i;
+	size_t i;
 
 	*repop = NULL;
 
@@ -694,7 +694,7 @@ const struct got_error *
 got_repo_close(struct got_repository *repo)
 {
 	const struct got_error *err = NULL, *child_err;
-	int i;
+	size_t i;
 
 	for (i = 0; i < nitems(repo->packidx_cache); i++) {
 		if (repo->packidx_cache[i] == NULL)
@@ -877,7 +877,7 @@ cache_packidx(struct got_repository *repo, struct got_packidx *packidx,
     const char *path_packidx)
 {
 	const struct got_error *err = NULL;
-	int i;
+	size_t i;
 
 	for (i = 0; i < nitems(repo->packidx_cache); i++) {
 		if (repo->packidx_cache[i] == NULL)
@@ -930,7 +930,7 @@ got_repo_search_packidx(struct got_packidx **packidx, int *idx,
 	DIR *packdir;
 	struct dirent *dent;
 	char *path_packidx;
-	int i;
+	size_t i;
 
 	/* Search pack index cache. */
 	for (i = 0; i < nitems(repo->packidx_cache); i++) {
@@ -1068,7 +1068,7 @@ got_repo_cache_pack(struct got_pack **packp, struct got_repository *repo,
 	const struct got_error *err = NULL;
 	struct got_pack *pack = NULL;
 	struct stat sb;
-	int i;
+	size_t i;
 
 	if (packp)
 		*packp = NULL;
@@ -1136,7 +1136,7 @@ struct got_pack *
 got_repo_get_cached_pack(struct got_repository *repo, const char *path_packfile)
 {
 	struct got_pack *pack = NULL;
-	int i;
+	size_t i;
 
 	for (i = 0; i < nitems(repo->packs); i++) {
 		pack = &repo->packs[i];
@@ -1166,7 +1166,7 @@ got_repo_init(const char *repo_path)
 	    "\tfilemode = true\n"
 	    "\tbare = true\n";
 	char *path;
-	int i;
+	size_t i;
 
 	if (!got_path_dir_is_empty(repo_path))
 		return got_error(GOT_ERR_DIR_NOT_EMPTY);
